+++ lib/Stroage/Exception/TaskException.php
@@ -1,0 +1,6 @@
+<?php
+namespace Stroage\Exception;
+
+class TaskException extends \Exception
+{
+}
